7-01-2026/requirements.txt
fastapi==0.115.9
uvicorn==0.30.6
chromadb==0.5.23
requests==2.32.5
pydantic==2.7.4
numpy==1.26.4
python-dotenv
